refactor(site/src/pages/AgentsPage): rename StickToBottom to ChatScrollContainer
Move ChatScrollContainer component from AgentChatPageView into its own
file alongside the useStickToBottom hook. The hook is now an internal
implementation detail of ChatScrollContainer.tsx.
